requirements.txt
langchain-google-genai
flask
python-dotenv
pypdf
langchain_community
faiss-cpu
langchain